+++ envs/container/docker/av-ffmpeg/alpine/avtool.sh
@@ -7,25 +7,9 @@
 #
 # executable
 ###########################################################################################################################################################################################
-set -u
-avdir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-cd "$avdir"
-
-# colors for formatting the ouput
-YELLOW='\033[1;33m'
-GREEN='\033[1;32m'
-RED='\033[0;31m'
-BLUE='\033[1;34m'
-NC='\033[0m' # No Color
-
-checkAVError() {
-    if [ $? -ne 0 ]; then
-        echo -e "\nAn error occured:
-    The script will be stopped."
-        exit 1
-    fi
-}
-
+set -eu
+repoRoot="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+cd "$repoRoot"
 #######################################################
 #- function used to print out script usage
 #######################################################
@@ -68,6 +52,20 @@ if [[ ! $action == login && ! $action == install && ! $action == start && ! $act
     usage
     exit 1
 fi
+# colors for formatting the ouput
+YELLOW='\033[1;33m'
+GREEN='\033[1;32m'
+RED='\033[0;31m'
+BLUE='\033[1;34m'
+NC='\033[0m' # No Color
+
+checkError() {
+    if [ $? -ne 0 ]; then
+        echo -e "${RED}\nAn error occured exiting from the current bash${NC}"
+        exit 1
+    fi
+}
+
 AV_SERVICE=av-ffmpeg
 AV_FLAVOR=alpine
 AV_IMAGE_NAME=${AV_SERVICE}-${AV_FLAVOR} 
@@ -76,8 +74,8 @@ AV_CONTAINER_NAME=av-ffmpeg-alpine-container
 AV_VOLUME=data1
 AV_FFMPEG_COMMAND="ffmpeg -y -nostats -loglevel 0  -i ./camera-300s.mkv -codec copy /${AV_VOLUME}/camera-300s.mp4"
 # Check if configuration file exists
-if [[ ! -f "$avdir"/"$configuration_file" ]]; then
-    cat > "$avdir"/"$configuration_file" << EOF
+if [[ ! -f "$repoRoot"/"$configuration_file" ]]; then
+    cat > "$repoRoot"/"$configuration_file" << EOF
 AV_IMAGE_NAME=${AV_IMAGE_NAME}
 AV_IMAGE_FOLDER=${AV_IMAGE_FOLDER}
 AV_CONTAINER_NAME=${AV_CONTAINER_NAME}
@@ -87,116 +85,97 @@ AV_TEMPDIR=$(mktemp -d)
 EOF
 fi
 # Read variables in configuration file
-export $(grep AV_IMAGE_NAME "$avdir"/"$configuration_file")
-export $(grep AV_IMAGE_FOLDER "$avdir"/"$configuration_file")
-export $(grep AV_CONTAINER_NAME "$avdir"/"$configuration_file")
-var=$(grep AV_FFMPEG_COMMAND "$avdir"/"$configuration_file")
+export $(grep AV_IMAGE_NAME "$repoRoot"/"$configuration_file")
+export $(grep AV_IMAGE_FOLDER "$repoRoot"/"$configuration_file")
+export $(grep AV_CONTAINER_NAME "$repoRoot"/"$configuration_file")
+var=$(grep AV_FFMPEG_COMMAND "$repoRoot"/"$configuration_file")
 cmd="export $var"
 eval $cmd
-export $(grep AV_VOLUME "$avdir"/"$configuration_file")
-export $(grep AV_TEMPDIR "$avdir"/"$configuration_file" |  { read test; if [[ -z $test ]] ; then AV_TEMPDIR=$(mktemp -d) ; echo "AV_TEMPDIR=$AV_TEMPDIR" ; echo "AV_TEMPDIR=$AV_TEMPDIR" >> .avtoolconfig ; else echo $test; fi } )
+export $(grep AV_VOLUME "$repoRoot"/"$configuration_file")
+export $(grep AV_TEMPDIR "$repoRoot"/"$configuration_file" |  { read test; if [[ -z $test ]] ; then AV_TEMPDIR=$(mktemp -d) ; echo "AV_TEMPDIR=$AV_TEMPDIR" ; echo "AV_TEMPDIR=$AV_TEMPDIR" >> .avtoolconfig ; else echo $test; fi } )
 
 if [[ -z "${AV_TEMPDIR}" ]] ; then
     AV_TEMPDIR=$(mktemp -d)
-    sed -i 's/AV_TEMPDIR=.*/AV_TEMPDIR=${AV_TEMPDIR}/' "$avdir"/"$configuration_file"
+    sed -i 's/AV_TEMPDIR=.*/AV_TEMPDIR=${AV_TEMPDIR}/' "$repoRoot"/"$configuration_file"
 fi
 
 if [[ "${action}" == "install" ]] ; then
     echo "Installing pre-requisite"
     echo "Installing azure cli"
     curl -sL https://aka.ms/InstallAzureCLIDeb | sudo bash
-    checkAVError
     echo "Installing ffmpeg"
     sudo apt-get -y update
     sudo apt-get -y install ffmpeg
-    checkAVError
-    echo "Downloading content"
-    wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv     
-    checkAVError
+    if [ ! -f "${AV_TEMPDIR}"/camera-300s.mkv ]; then
+        echo "Downloading content"
+        wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv     
+    fi
     echo "Installing docker"
     # removing old version
     sudo apt-get remove docker docker-engine docker.io containerd runc
-    checkAVError
     sudo apt-get -y update
-    checkAVError
     sudo apt-get -y install \
         apt-transport-https \
         ca-certificates \
         curl \
         gnupg-agent \
         software-properties-common
-    checkAVError
     curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-    checkAVError
     sudo apt-key fingerprint 0EBFCD88
-    checkAVError
     sudo add-apt-repository \
-        "deb [ar
-        ch=amd64] https://download.docker.com/linux/ubuntu \
+        "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
         $(lsb_release -cs) \
         stable"
     sudo apt-get -y update
-    checkAVError
     sudo apt-get -y install  jq
     sudo apt-get -y install docker.io    
-    checkAVError
     sudo groupadd docker || true
     sudo usermod -aG docker ${USER} || true
-    checkAVError
+    echo -e "${GREEN}Installing pre-requisites done${NC}"
     exit 0
 fi
 if [[ "${action}" == "login" ]] ; then
     echo "Login..."
     docker login
-    checkAVError
-    echo "Login done"
+    echo -e "${GREEN}Login done${NC}"
     exit 0
 fi
 
 if [[ "${action}" == "deploy" ]] ; then
     echo "Deploying service..."
-    sudo docker container stop ${AV_CONTAINER_NAME} > /dev/null 2> /dev/null  || true
     sudo docker container rm ${AV_CONTAINER_NAME} > /dev/null 2> /dev/null  || true
-    sudo docker image rm ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} > /dev/null 2> /dev/null  || true 
+    sudo docker image rm ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} > /dev/null 2> /dev/null  || true
     sudo docker build -t ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} .
-    checkAVError
-    sudo docker run -d  -it -v ${AV_TEMPDIR}:/${AV_VOLUME} --name ${AV_CONTAINER_NAME} ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} ${AV_FFMPEG_COMMAND}
-    checkAVError
-    echo "Deployment done"
+    sudo docker run  -d -it -v ${AV_TEMPDIR}:/${AV_VOLUME} --name ${AV_CONTAINER_NAME} ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} ${AV_FFMPEG_COMMAND}
+    echo -e "${GREEN}Deployment done${NC}"
     exit 0
 fi
 
 if [[ "${action}" == "undeploy" ]] ; then
     echo "Undeploying service..."
-    sudo docker container stop ${AV_CONTAINER_NAME} > /dev/null 2> /dev/null  || true
     sudo docker container rm ${AV_CONTAINER_NAME} > /dev/null 2> /dev/null  || true
-    checkAVError
     sudo docker image rm ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} > /dev/null 2> /dev/null  || true
-    checkAVError
-    echo "Undeployment done"
+    echo -e "${GREEN}Undeployment done${NC}"
     exit 0
 fi
 if [[ "${action}" == "status" ]] ; then
     echo "Checking status..."
     sudo docker container inspect ${AV_CONTAINER_NAME} --format '{{json .State.Status}}'
-    checkAVError
-    echo "Status done"
+    echo -e "${GREEN}Container status done${NC}"
     exit 0
 fi
 
 if [[ "${action}" == "start" ]] ; then
     echo "Starting service..."
     sudo docker container start ${AV_CONTAINER_NAME}
-    checkAVError
-    echo "Start done"
+    echo -e "${GREEN}Container started${NC}"
     exit 0
 fi
 
 if [[ "${action}" == "stop" ]] ; then
     echo "Stopping service..."
     sudo docker container stop ${AV_CONTAINER_NAME}
-    checkAVError
-    echo "Stop done"
+    echo -e "${GREEN}Container stopped${NC}"
     exit 0
 fi
 if [[ "${action}" == "test" ]] ; then
@@ -211,7 +190,6 @@ if [[ "${action}" == "test" ]] ; then
         exit 1
     fi
     sudo docker container start -i ${AV_CONTAINER_NAME}
-    checkAVError
     echo "Output directory : ${AV_TEMPDIR}"
     if [[ ! -f "${AV_TEMPDIR}/camera-300s.mp4" ]] ; then
         echo "ffmpeg Test failed - check file ${AV_TEMPDIR}/camera-300s.mp4"
@@ -220,7 +198,7 @@ if [[ "${action}" == "test" ]] ; then
     fi
     echo "File ${AV_TEMPDIR}/camera-300s.mp4 exists"
     echo "Testing ffmpeg successful"
-    echo "TESTS SUCCESSFUL"
     sudo docker container stop ${AV_CONTAINER_NAME} > /dev/null 2> /dev/null  || true
+    echo -e "${GREEN}TESTS SUCCESSFUL${NC}"
     exit 0
 fi
