+++ envs/container/docker/av-ffmpeg/ubuntu/avtool.sh
@@ -52,6 +52,20 @@ if [[ ! $action == login && ! $action == install && ! $action == start && ! $act
     usage
     exit 1
 fi
+# colors for formatting the ouput
+YELLOW='\033[1;33m'
+GREEN='\033[1;32m'
+RED='\033[0;31m'
+BLUE='\033[1;34m'
+NC='\033[0m' # No Color
+
+checkError() {
+    if [ $? -ne 0 ]; then
+        echo -e "${RED}\nAn error occured exiting from the current bash${NC}"
+        exit 1
+    fi
+}
+
 AV_SERVICE=av-ffmpeg
 AV_FLAVOR=ubuntu
 AV_IMAGE_NAME=${AV_SERVICE}-${AV_FLAVOR} 
@@ -92,8 +106,10 @@ if [[ "${action}" == "install" ]] ; then
     echo "Installing ffmpeg"
     sudo apt-get -y update
     sudo apt-get -y install ffmpeg
-    echo "Downloading content"
-    wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv     
+    if [ ! -f "${AV_TEMPDIR}"/camera-300s.mkv ]; then
+        echo "Downloading content"
+        wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv     
+    fi
     echo "Installing docker"
     # removing old version
     sudo apt-get remove docker docker-engine docker.io containerd runc
@@ -115,12 +131,13 @@ if [[ "${action}" == "install" ]] ; then
     sudo apt-get -y install docker.io    
     sudo groupadd docker || true
     sudo usermod -aG docker ${USER} || true
+    echo -e "${GREEN}Installing pre-requisites done${NC}"
     exit 0
 fi
 if [[ "${action}" == "login" ]] ; then
     echo "Login..."
     docker login
-    echo "Login done"
+    echo -e "${GREEN}Login done${NC}"
     exit 0
 fi
 
@@ -130,7 +147,7 @@ if [[ "${action}" == "deploy" ]] ; then
     sudo docker image rm ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} > /dev/null 2> /dev/null  || true
     sudo docker build -t ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} .
     sudo docker run  -d -it -v ${AV_TEMPDIR}:/${AV_VOLUME} --name ${AV_CONTAINER_NAME} ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} ${AV_FFMPEG_COMMAND}
-    echo "Deployment done"
+    echo -e "${GREEN}Deployment done${NC}"
     exit 0
 fi
 
@@ -138,27 +155,27 @@ if [[ "${action}" == "undeploy" ]] ; then
     echo "Undeploying service..."
     sudo docker container rm ${AV_CONTAINER_NAME} > /dev/null 2> /dev/null  || true
     sudo docker image rm ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} > /dev/null 2> /dev/null  || true
-    echo "Undeployment done"
+    echo -e "${GREEN}Undeployment done${NC}"
     exit 0
 fi
 if [[ "${action}" == "status" ]] ; then
     echo "Checking status..."
     sudo docker container inspect ${AV_CONTAINER_NAME} --format '{{json .State.Status}}'
-    echo "Status done"
+    echo -e "${GREEN}Container status done${NC}"
     exit 0
 fi
 
 if [[ "${action}" == "start" ]] ; then
     echo "Starting service..."
     sudo docker container start ${AV_CONTAINER_NAME}
-    echo "Start done"
+    echo -e "${GREEN}Container started${NC}"
     exit 0
 fi
 
 if [[ "${action}" == "stop" ]] ; then
     echo "Stopping service..."
     sudo docker container stop ${AV_CONTAINER_NAME}
-    echo "Stop done"
+    echo -e "${GREEN}Container stopped${NC}"
     exit 0
 fi
 if [[ "${action}" == "test" ]] ; then
@@ -181,7 +198,7 @@ if [[ "${action}" == "test" ]] ; then
     fi
     echo "File ${AV_TEMPDIR}/camera-300s.mp4 exists"
     echo "Testing ffmpeg successful"
-    echo "TESTS SUCCESSFUL"
     sudo docker container stop ${AV_CONTAINER_NAME} > /dev/null 2> /dev/null  || true
+    echo -e "${GREEN}TESTS SUCCESSFUL${NC}"
     exit 0
 fi
