+++ envs/container/docker/av-rtmp-sink/alpine/avtool.sh
@@ -51,6 +51,20 @@ if [[ ! $action == login && ! $action == install && ! $action == start && ! $act
     usage
     exit 1
 fi
+# colors for formatting the ouput
+YELLOW='\033[1;33m'
+GREEN='\033[1;32m'
+RED='\033[0;31m'
+BLUE='\033[1;34m'
+NC='\033[0m' # No Color
+
+checkError() {
+    if [ $? -ne 0 ]; then
+        echo -e "${RED}\nAn error occured exiting from the current bash${NC}"
+        exit 1
+    fi
+}
+
 AV_SERVICE=av-rtmp-sink
 AV_FLAVOR=alpine
 AV_IMAGE_NAME=${AV_SERVICE}-${AV_FLAVOR} 
@@ -101,8 +115,10 @@ if [[ "${action}" == "install" ]] ; then
     echo "Installing ffmpeg"
     sudo apt-get -y update
     sudo apt-get -y install ffmpeg
-    echo "Downloading content"
-    wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv     
+    if [ ! -f "${AV_TEMPDIR}"/camera-300s.mkv ]; then
+        echo "Downloading content"
+        wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv     
+    fi
     echo "Installing docker"
     # removing old version
     sudo apt-get remove docker docker-engine docker.io containerd runc
@@ -124,12 +140,13 @@ if [[ "${action}" == "install" ]] ; then
     sudo apt-get -y install docker.io    
     sudo groupadd docker || true
     sudo usermod -aG docker ${USER} || true
+    echo -e "${GREEN}Installing pre-requisites done${NC}"
     exit 0
 fi
 if [[ "${action}" == "login" ]] ; then
     echo "Login..."
     docker login
-    echo "Login done"
+    echo -e "${GREEN}Login done${NC}"
     exit 0
 fi
 
@@ -140,7 +157,7 @@ if [[ "${action}" == "deploy" ]] ; then
     sudo docker image rm ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} > /dev/null 2> /dev/null  || true
     sudo docker build  --build-arg  AV_PORT_RTMP=${AV_PORT_RTMP} --build-arg  AV_PORT_SSL=${AV_PORT_SSL} --build-arg  AV_PORT_HTTP=${AV_PORT_HTTP} --build-arg  AV_PORT_HLS=${AV_PORT_HLS}  --build-arg  AV_HOSTNAME=${AV_HOSTNAME} --build-arg  AV_COMPANYNAME=${AV_COMPANYNAME} -t ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} . 
     sudo docker run  -d -it -p ${AV_PORT_HTTP}:${AV_PORT_HTTP}/tcp  -p ${AV_PORT_HLS}:${AV_PORT_HLS}/tcp    -p ${AV_PORT_RTMP}:${AV_PORT_RTMP}/tcp   -p ${AV_PORT_SSL}:${AV_PORT_SSL}/tcp  -e PORT_RTMP=${AV_PORT_RTMP} -e PORT_SSL=${AV_PORT_SSL} -e PORT_HTTP=${AV_PORT_HTTP} -e PORT_HLS=${AV_PORT_HLS}  -e HOSTNAME=${AV_HOSTNAME} -e COMPANYNAME=${AV_COMPANYNAME} --name ${AV_CONTAINER_NAME} ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} 
-    echo "Deployment done"
+    echo -e "${GREEN}Deployment done${NC}"
     exit 0
 fi
 
@@ -149,34 +166,36 @@ if [[ "${action}" == "undeploy" ]] ; then
     sudo docker container stop ${AV_CONTAINER_NAME} > /dev/null 2> /dev/null  || true
     sudo docker container rm ${AV_CONTAINER_NAME} > /dev/null 2> /dev/null  || true
     sudo docker image rm ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} > /dev/null 2> /dev/null  || true
-    echo "Undeployment done"
+    echo -e "${GREEN}Undeployment done${NC}"
     exit 0
 fi
 if [[ "${action}" == "status" ]] ; then
     echo "Checking status..."
     sudo docker container inspect ${AV_CONTAINER_NAME} --format '{{json .State.Status}}'
-    echo "Status done"
+    echo -e "${GREEN}Container status done${NC}"
     exit 0
 fi
 
 if [[ "${action}" == "start" ]] ; then
     echo "Starting service..."
     sudo docker container start ${AV_CONTAINER_NAME}
-    echo "Start done"
+    echo -e "${GREEN}Container started${NC}"
     exit 0
 fi
 
 if [[ "${action}" == "stop" ]] ; then
     echo "Stopping service..."
     sudo docker container stop ${AV_CONTAINER_NAME}
-    echo "Stop done"
+    echo -e "${GREEN}Container stopped${NC}"
     exit 0
 fi
 if [[ "${action}" == "test" ]] ; then
     rm -f "${AV_TEMPDIR}"/*.mp4
     rm -f "${AV_TEMPDIR}"/*.mkv
-    echo "Downloading content"
-    wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv 
+    if [ ! -f "${AV_TEMPDIR}"/camera-300s.mkv ]; then
+        echo "Downloading content"
+        wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv     
+    fi
     echo "Start ${AV_CONTAINER_NAME} container..."
     sudo docker container start ${AV_CONTAINER_NAME}
 
@@ -207,8 +226,8 @@ if [[ "${action}" == "test" ]] ; then
         exit 1
     fi    
     echo "Testing ${AV_CONTAINER_NAME} successful"
-    echo "TESTS SUCCESSFUL"
     kill %1
     sudo docker container stop ${AV_CONTAINER_NAME} > /dev/null 2> /dev/null  || true
+    echo -e "${GREEN}TESTS SUCCESSFUL${NC}"
     exit 0
 fi
